Fix: CKeyBind.cppでDebug_x64のときCStrictIntegerのエラー
[patchunicode:#862]


git-svn-id: https://svn.code.sf.net/p/sakura-editor/code/sakura/trunk2@3770 f7ce1907-e4c7-47ca-9f76-12c87ed2c91c

+++ sakura_core/func/CKeyBind.cpp
@@ -487,7 +487,7 @@ TCHAR* CKeyBind::GetMenuLabel(
 		/* 機能に対応するキー名の取得 */
 		if( GetKeyStr( hInstance, nKeyNameArrNum, pKeyNameArr, cMemAccessKey, nFuncId, bGetDefFuncCode ) ){
 			// バッファが足りないときは入れない
-			if( _tcslen( pszLabel ) + cMemAccessKey.GetStringLength() + 1 < LABEL_MAX ){
+			if( _tcslen( pszLabel ) + (Int)cMemAccessKey.GetStringLength() + 1 < LABEL_MAX ){
 				_tcscat( pszLabel, _T("\t") );
 				_tcscat( pszLabel, cMemAccessKey.GetStringPtr() );
 			}
